Finished pay and fixed .cgi labels
did the pay by checking if the order was already in the table pay and if the customer associated with the order is the same as the one that's paying. If neither of these errors happen the order just gets added to the table pay.

Also fixed the lables on the .cgi's for change, pay and the removes, so that the know what forms to grab from, but im not sure the thing i wanted was the form lable, or the form name.
Also changed the forms and HTML's where there is app.cgi to the respective .cgi